add columnNames to table.props / refactor postDataMigration
