+++ test/dataMigration/types.ts
@@ -1,0 +1,3 @@
+import { NodeValue } from '@meta/assessment'
+
+export type NodeRow = { country_iso: string; row_uuid: string; col_uuid: string; value: NodeValue }
